@@ -31,20 +31,18 @@
     "install": "^0.13.0",
     "md5": "^2.3.0",
     "multiparty": "^4.2.3",
-    "next": "^12.1.6",
+    "next": "^13.1.1",
     "node-fetch": "^3.2.10",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-error-boundary": "^3.1.4",
-    "react-relay": "^13.0.2",
+    "react-relay": "^0.0.0-main-83ae9735",
     "react-virtuoso": "^2.18.0",
-    "relay-nextjs": "^0.6.0",
-    "relay-runtime": "^13.0.2",
+    "relay-runtime": "^0.0.0-main-83ae9735",
     "simple-oauth2": "^4.3.0",
     "string-strip-html": "^9.1.12"
   },
   "devDependencies": {
-    "@babel/core": "^7.18.6",
     "@storybook/addon-actions": "^6.5.9",
     "@storybook/addon-essentials": "^6.5.9",
     "@storybook/addon-interactions": "^6.5.9",
@@ -60,16 +58,14 @@
     "@types/multiparty": "^0.0.33",
     "@types/node": "17.0.15",
     "@types/react": "17.0.39",
-    "@types/react-relay": "^13.0.1",
-    "@types/relay-runtime": "^13.0.1",
+    "@types/react-relay": "^14.1.2",
+    "@types/relay-runtime": "14.1.4",
     "@types/simple-oauth2": "^4.1.1",
     "@typescript-eslint/eslint-plugin": "^5.28.0",
     "@typescript-eslint/parser": "^5.28.0",
     "autoprefixer": "^10.4.2",
-    "babel-loader": "^8.2.5",
-    "babel-plugin-relay": "^13.0.2",
     "eslint": "^8.18.0",
-    "eslint-config-next": "12.0.10",
+    "eslint-config-next": "^13.1.1",
     "eslint-config-prettier": "^8.5.0",
     "eslint-import-resolver-typescript": "^2.7.1",
     "eslint-plugin-prettier": "^4.0.0",
@@ -80,9 +76,7 @@
     "lint-staged": "^13.0.2",
     "postcss": "^8.4.6",
     "prettier": "^2.7.1",
-    "relay-compiler": "^13.0.2",
-    "relay-compiler-language-typescript": "^15.0.1",
-    "relay-config": "^12.0.1",
+    "relay-compiler": "^0.0.0-main-83ae9735",
     "relay-test-utils": "^14.0.0",
     "storybook-css-modules-preset": "^1.1.1",
     "tailwind-scrollbar": "^2.0.1",
