@@ -7,7 +7,16 @@ const nextConfig = {
   //   }
 
   //   return config;
-  // },
+  // },  reactStrictMode: true,
+  swcMinify: true,
+  compiler: {
+    relay: {
+      src: './',
+      language: 'typescript',
+      artifactDirectory: '__generated__',
+    },
+  },
+  experimental: { appDir: true },
   async rewrites() {
     return [
       {
