@@ -1,7 +1,7 @@
 import dayjs from 'dayjs'
 import relativeTime from 'dayjs/plugin/relativeTime'
 import { graphql, useFragment } from 'react-relay'
-import { Comment_node$key } from '../../queries/__generated__/Comment_node.graphql'
+import { Comment_node$key } from '__generated__/Comment_node.graphql'
 
 type CommentProps = {
   comment: Comment_node$key
