@@ -1,5 +1,5 @@
 import { graphql, useFragment } from 'react-relay'
-import { StoryPreview_node$key } from '../../queries/__generated__/StoryPreview_node.graphql'
+import { StoryPreview_node$key } from '__generated__/StoryPreview_node.graphql'
 import StoryPreviewFooter from './StoryPreviewFooter'
 import UserSnippet from './UserSnippet'
 import { GoKebabVertical } from '@react-icons/all-files/go/GoKebabVertical'
@@ -34,25 +34,23 @@ const StoryPreview = ({ story }: StoryPreviewProps) => {
     <ErrorBoundary FallbackComponent={ErrorFallback}>
       <div className="my-8 mr-2 flex flex-col bg-white  dark:bg-dark-700 rounded-md">
         <Link href={`/story/${data.id}___${data.urlSuffix}`}>
-          <a>
-            <div>
-              {data.thumbnail ? (
-                <img className="cover-image" src={data.thumbnail} alt="" />
-              ) : (
-                <div className="image-placeholder bg-gray-100 dark:bg-dark-500"></div>
-              )}
+          <div>
+            {data.thumbnail ? (
+              <img className="cover-image" src={data.thumbnail} alt="" />
+            ) : (
+              <div className="image-placeholder bg-gray-100 dark:bg-dark-500" />
+            )}
+          </div>
+          <div className="px-4 pb-4">
+            <div className="flex flex-row justify-between">
+              <UserSnippet user={data.createdBy} />
+              <GoKebabVertical className="mt-4" />
             </div>
-            <div className="px-4 pb-4">
-              <div className="flex flex-row justify-between">
-                <UserSnippet user={data.createdBy} />
-                <GoKebabVertical className="mt-4" />
-              </div>
-              <div className="prose dark:text-gray-100">
-                <h3 className="dark:text-gray-200">{data.title}</h3>
-                <p className="abstract-content">{data.abstractContent}</p>
-              </div>
+            <div className="prose dark:text-gray-100">
+              <h3 className="dark:text-gray-200">{data.title}</h3>
+              <p className="abstract-content">{data.abstractContent}</p>
             </div>
-          </a>
+          </div>
         </Link>
         <StoryPreviewFooter story={data} />
         <style jsx>
